pyproject.toml
[project]
name = "mailmerge"
version = "0.1.0"
description = """
Email data collection and aggregation system for academic departments
"""
authors = [
    {name = "Dubley", email = "2590684243@qq.com"}
]
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "psycopg2-binary>=2.9.9",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "python-multipart>=0.0.6",
    "openpyxl>=3.1.2",
    "pandas>=2.1.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.11.0",
    "flake8>=6.1.0",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 88
target-version = ['py39']